logout

+++ test/client/session_test.rb
@@ -6,7 +6,7 @@ class SessionTest < MiniTest::Test
   SESSION_FILE = __dir__ + '/session'
 
   def teardown
-    File.delete(SESSION_FILE)
+    File.delete(SESSION_FILE) if File.exist?(SESSION_FILE)
   end
 
   def test_immutable_token
@@ -23,9 +23,22 @@ class SessionTest < MiniTest::Test
     assert(false == File.exist?(SESSION_FILE))
 
     session.token = "Test1"
+    assert_equal("Test1", session.token)
     assert(File.exist?(SESSION_FILE))
 
     session2 = Client::Session.new(__dir__)
     assert_equal("Test1", session2.token)
+  end
+
+  def test_clear
+    session = Client::Session.new(__dir__)
+
+    session.token = "Test1"
+    assert_equal("Test1", session.token)
+    assert(File.exist?(SESSION_FILE))
+
+    session.clear
+    assert(false == File.exist?(SESSION_FILE))
+    assert_nil(session.token)
   end
 end
